Added an arrow and slowdown while it's spinning
Decided to take a break from text stuff and get the core functionality of the wheel to work. The wheel will now slow down while spinning, and there's an arrow to point at a slice once the wheel has stopped. This means you can technically use this app to select stuff now