SQL database elements have been added
Items on the wheel are now stored and loaded from a local SQLite database. This means that items added to the table will not go away upon closing the program. Additions, deletions, and loads all seem to working from the tests I ran, but I will need to make sure through further testing.

With this feature being added, the program is more or less "done". There is still work to do, bugs to fix, and features to polish, but the wheel now has all the functionality I was hoping to give it.