Working on making it so that you can resize the window.
Did some small edits so that window resizing is allowed and everything remains at the position they should be at. The only thing that actually gets bigger though is the wheel itself (not including the section text within). I'll have to figure out a clean way to do that at some point.