Minor cleanup
I thought that my hardcoding of "Comic Sans MS" as the font would lead to issues, so I came back to fix it. Turns out, it wasn't an issue at all and that pygame automatically just chooses the default system font if it can't find the specified one.

I decided to do some minor cleanup on the code though while I was back.